move docker api code
move docker api code out of index.js into lib/docker
