+++ index.js
@@ -6,11 +6,11 @@ const express = require('express')
 const bodyParser = require('body-parser')
 const app = express()
 const config = require('./lib/config')
-const child_process = require('child_process')
+const docker = require('./lib/docker')
 const logger = require('./lib/logger')('DOCKER-DEPLOY')
 const services = require(`./config/config.json`)[config.whichConfig]
 
-const { dockerCommand, port, token, username, password, withRegistryAuth } = config
+const { port, token } = config
 
 app.use(bodyParser.json())
 app.use(bodyParser.urlencoded({ extended: true }))
@@ -25,30 +25,13 @@ app.post('/webhook/:token', (req, res) => {
   res.send('OK')
 
   const payload = req.body
-  const image = `${payload.repository.repo_name}:${payload.push_data.tag}`
+  const imageName = `${payload.repository.repo_name}:${payload.push_data.tag}`
 
-  if (!services[image]) return logger.log(`Received updated for "${image}" but not configured to handle updates for this image.`)
+  if (!services[imageName]) return logger.log(`Received updated for "${imageName}" but not configured to handle updates for this image.`)
 
-  const service = services[image].service
-
-  // Make sure we are logged in to be able to pull the image
-  child_process.exec(`${dockerCommand} login -u "${username}" -p "${password}"`,
-    (error, stdout, stderr) => {
-      if (error) return logger.error(error)
-
-      // Deploy the image and force a restart of the associated service
-      logger.log(`Deploying ${image} to ${service}…`)
-
-      const registryAuth = (withRegistryAuth) ? '--with-registry-auth' : ''
-      child_process.exec(`${dockerCommand} service update ${service} --force ${registryAuth} --image=${image}`,
-        (error, stdout, stderr) => {
-          if (error) {
-            logger.error(`Failed to deploy ${image} to ${service}!`)
-            return logger.error(error)
-          }
-          logger.log(`Deployed ${image} to ${service} successfully and restarted the service.`)
-        })
-    })
+  const image = services[imageName]
+  const service = image.service
+  docker.deploy(imageName, service)
 })
 
 app.all('*', (req, res) => {
